Feature/next (#17)
* Make The recipe listing a seperate component

* allow to delete a recipe

* present confirmation alert when deleting a recipe

+++ src/app/listings-page/listings-page.component.ts
@@ -67,15 +67,16 @@ export class ListingsPageComponent implements OnInit {
   }
 
   public onDeleteRecipeClicked(recipe: Recipe) {
-    // TOOD: we need to stopPropegation on the event to prevent also clike on the recipe clicked. This means the entire recipe should be a component with outputs, so the click will only pass the event and not the recipe as now.
-    this.utilsService.deleteFromFirebaseCloudFireStoreRecipeID(recipe.id).then((deletionSuccess) => {
-      if (deletionSuccess) {
-        this.updateRecipes();
-      } else {
-        console.log('deletion failure');
-      }
-
-    });
+    let proceed = confirm("Are you sure you want to delete " + '"' + recipe.title + '"?');
+    if (proceed) {
+      this.utilsService.deleteFromFirebaseCloudFireStoreRecipeID(recipe.id).then((deletionSuccess) => {
+        if (deletionSuccess) {
+          this.updateRecipes();
+        } else {
+          console.log('deletion failure');
+        }
+      });
+    } 
   }
 
   public signOutBtnClicked(){
